examples: Refactor command-line parsing
Move some commonly used command-line flag parsing code into
e2d2::config::flag_reader.
